pyproject.toml
[project]
name = "geortacking-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "folium>=0.20.0",
    "kafka-python>=2.2.15",
    "pymongo>=4.14.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.35.0",
]
